+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/dto/review/ReviewRespDto.java
@@ -8,6 +8,7 @@ import org.springframework.data.domain.Page;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Map;
 
 public class ReviewRespDto {
 
@@ -32,5 +33,26 @@ public class ReviewRespDto {
             this.createdReviewDate = createdReviewDate;
             this.content = content;
         }
+    }
+
+    @AllArgsConstructor
+    @Data
+    public static class MyReviewRespDto {
+
+        private Long reviewId;
+        private Integer rating;
+        @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
+        private LocalDateTime createdReviewDate;
+        private String content;
+        private List<String> reviewImages;
+    }
+
+    @AllArgsConstructor
+    @Data
+    public static class WaterPlaceReviewDetailRespDto {
+        private Double waterPlaceRating; // 물놀이 장소 평점
+        private int reviewCnt; // 리뷰 수
+        private Map<Integer, Long> ratingRatio;
+        private Page<ReviewRespDto.WaterPlaceReviewRespDto> reviews; // 리뷰
     }
 }
