+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/service/accident/AccidentService.java
@@ -1,6 +1,7 @@
 package kr.ac.kumoh.illdang100.tovalley.service.accident;
 
-import static kr.ac.kumoh.illdang100.tovalley.dto.page.PageRespDto.*;
+
+import static kr.ac.kumoh.illdang100.tovalley.dto.accident.AccidentRespDto.*;
 
 public interface AccidentService {
