+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/service/page/PageService.java
@@ -8,7 +8,7 @@ public interface PageService {
 
     MainPageAllRespDto getMainPageAllData();
 
-    WaterPlaceDetailAllRespDto getWaterPlaceDetailPageAllData(Long waterPlaceId, Pageable pageable);
+    WaterPlaceDetailPageAllRespDto getWaterPlaceDetailPageAllData(Long waterPlaceId, Pageable pageable);
 
     MyPageAllRespDto getMyPageAllData();
 }
